requirements.txt
mcp>=1.0.0
google-ads>=24.0.0
facebook-business>=20.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
